Separate out document conversion from scanning.
